プロセスが生存している場合に終了させる操作をする

+++ lib763/multp/multp.py
@@ -25,7 +25,8 @@ def stop_process(process: Process) -> None:
     Parameters:
     process (Process): 終了させるプロセスオブジェクト
     """
-    process.terminate()
+    if process.is_alive():
+        process.terminate()
 
 
 def parallel_process(func: Callable[[Any], Any], data_list: List[Any]) -> List[Any]:
